requirements.txt
argparse
matplotlib
numpy
Pillow
statistics
torch
torchnet
torchvision
tqdm
